Started connecting backend to screens

+++ app/src/main/java/com/example/studyflow/dao/TimersessionDao.kt
@@ -8,10 +8,10 @@ import com.example.studyflow.model.Timersession
 @Dao
 interface TimersessionDao : BaseDao<Timersession> {
     @Query("SELECT * FROM timersessions WHERE userId = :userId")
-    suspend fun getTimersessionsByUserId(userId: Int): List<Timersession>
+    suspend fun getTimersessionsByUserId(userId: Long): List<Timersession>
 
     @Query("SELECT * FROM timersessions WHERE userId = :userId AND timerType = :timerType")
-    suspend fun getTimerSessionsByUserIdAndTimerType(userId: Int, timerType: TimerType): List<Timersession>
+    suspend fun getTimerSessionsByUserIdAndTimerType(userId: Long, timerType: TimerType): List<Timersession>
 
     // get by timer type
     // get by date or something that orders/groups by date
